chore(suite): staking rewards frequency & next rewards
