+++ packages/suite/src/views/wallet/staking/components/StakingDashboard/components/PayoutCardNextRewards.tsx
@@ -1,0 +1,60 @@
+import { useMemo } from 'react';
+
+import { BACKUP_REWARD_PAYOUT_DAYS } from '@suite-common/wallet-constants';
+import { getStakingDataForNetwork } from '@suite-common/wallet-utils';
+import { Paragraph } from '@trezor/components';
+import { BigNumber } from '@trezor/utils/src/bigNumber';
+
+import { Translation } from 'src/components/suite/Translation';
+import { useSelector } from 'src/hooks/suite';
+import { selectSelectedAccount } from 'src/reducers/wallet/selectedAccountReducer';
+
+import { PayoutCard } from './PayoutCard';
+
+interface PayoutCardNextRewardsProps {
+    nextRewardPayout?: number | null;
+    daysToAddToPool?: number;
+    validatorWithdrawTime?: number;
+}
+
+export const PayoutCardNextRewards = ({
+    nextRewardPayout,
+    daysToAddToPool,
+    validatorWithdrawTime,
+}: PayoutCardNextRewardsProps) => {
+    const selectedAccount = useSelector(selectSelectedAccount);
+
+    const { autocompoundBalance = '0' } = getStakingDataForNetwork(selectedAccount) ?? {};
+
+    const payout = useMemo(() => {
+        if (!nextRewardPayout || !daysToAddToPool) return undefined;
+
+        if (new BigNumber(autocompoundBalance).gt(0) || daysToAddToPool <= nextRewardPayout) {
+            return nextRewardPayout;
+        }
+
+        if (!validatorWithdrawTime) return undefined;
+
+        return Math.round(validatorWithdrawTime / 60 / 60 / 24) + nextRewardPayout;
+    }, [autocompoundBalance, daysToAddToPool, nextRewardPayout, validatorWithdrawTime]);
+
+    return (
+        <PayoutCard>
+            <>
+                <Paragraph typographyStyle="titleMedium">
+                    {payout === undefined ? (
+                        <Translation
+                            id="TR_STAKE_MAX_REWARD_DAYS"
+                            values={{ count: BACKUP_REWARD_PAYOUT_DAYS }}
+                        />
+                    ) : (
+                        <Translation id="TR_STAKE_DAYS" values={{ count: payout }} />
+                    )}
+                </Paragraph>
+                <Paragraph typographyStyle="hint" variant="tertiary">
+                    <Translation id="TR_STAKE_NEXT_PAYOUT" />
+                </Paragraph>
+            </>
+        </PayoutCard>
+    );
+};
